+++ src/components/Icon/index.tsx
@@ -2,6 +2,7 @@ import React, {FC} from 'react';
 import {createIconSetFromIcoMoon} from 'react-native-vector-icons';
 import icoMoonConfig from './assets/fonts/selection.json';
 import {primary} from '../../theme/palette';
+import {moderateScale, scaledForDevice} from '../../scale';
 
 const IconComponent = createIconSetFromIcoMoon(
 	icoMoonConfig,
@@ -20,7 +21,9 @@ const Icon: FC<Props> = ({name, color = primary.main, size = 16, ...props}) => {
 	if (!name) {
 		return null;
 	}
-	return <IconComponent name={name} color={color} size={size} {...props} />;
+	const validateSize = scaledForDevice(size, moderateScale);
+
+	return <IconComponent name={name} color={color} size={validateSize} {...props} />;
 };
 
 export default Icon;