+++ src/components/Select/Components/Icons/Chevron/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import {Pressable, View, ViewStyle} from 'react-native';
 import {primary} from '../../../../../theme/palette';
 import Svg, {Path} from 'react-native-svg';
+import {moderateScale, scaledForDevice} from '../../../../../scale';
 
 interface IconProps {
 	color?: string;
@@ -11,11 +12,13 @@ interface IconProps {
 	onPress?: () => void;
 }
 
-const Chevron = ({style, color, size, onPress, ...props}: IconProps) => {
+const Chevron = ({style, color, size = 21, onPress, ...props}: IconProps) => {
+	const validSize = scaledForDevice(size, moderateScale);
+
 	return (
 		<Pressable onPress={onPress} style={style} {...props}>
 			<View>
-				<Svg width={size} height={size} viewBox="0 0 16 16">
+				<Svg width={validSize} height={validSize} viewBox="0 0 16 16">
 					<Path
 						d="M8.17432 11.1055L3 6.49316L4.33106 5L8.17627 8.42773L12.0132 5.01904L13.3413 6.51416L8.17432 11.1055Z"
 						fill={color}
@@ -28,7 +31,6 @@ const Chevron = ({style, color, size, onPress, ...props}: IconProps) => {
 
 Chevron.defaultProps = {
 	color: primary.main,
-	size: 21,
 	onPress: () => {},
 };