+++ src/components/Svg/index.tsx
@@ -7,6 +7,7 @@ import JanisLogo from './svgs/JanisLogo';
 import JanisLogoColor from './svgs/JanisLogoColor';
 import LoginIllustration from './svgs/LoginIllustration';
 import NoNotifications from './svgs/NoNotifications';
+import {horizontalScale, moderateScale, scaledForDevice} from '../../scale';
 
 const svgs = {
 	'empty-illustration': EmptyIllustration,
@@ -22,9 +23,17 @@ const Svg = ({name, width, height, size, ...props}: Isvg) => {
 		return null;
 	}
 	const SvgSelected = svgs[name];
+	const selectedWidth = size ?? width;
+	const selectedHeight = size ?? height;
+	const parseSelectedWidth = selectedWidth || 0;
+	const parseSelectedHeight = selectedHeight || 0;
+
+	const validateWidth = scaledForDevice(parseSelectedWidth, horizontalScale);
+	const validateHeight = scaledForDevice(parseSelectedHeight, moderateScale);
+
 	return (
 		<View {...props}>
-			<SvgSelected width={size ?? width} height={size ?? height} {...props} />
+			<SvgSelected width={validateWidth} height={validateHeight} {...props} />
 		</View>
 	);
 };